Fix failing triangulation tests due to inlier flag
Creating a projected track set now marks the projected tracks
as all inliers.  It's up to individual tests to now clear all
the inlier flags if they want to pretend that these are not
"ground truth" tracks.  This fixes an error where the ground
truth projected tracks were all consider outliers and we could
not measure the inlier RMSE error.