Update the KPF YAML parser to allow embedded YAML in metadata.
Some of the KPF coming out of the annotation toolchain contains
embedded YAML in the metadata, e.g.

   - { meta: { team: "Kitware" } }

...previously, the parser attempted to render the metadata as a
string via the as<string>() method. This fails for YAML objects, which
require a YAML::Emitter. So we first try the as<string>() approach; if
this fails, we use an emitter, which renders the above as

    meta '{team: !<!> Kitware}'

This parses to itself if used as input.