Generalize bindings for other algorithms
The bindings follow the pattern that is similar to pybind11 tests. Thus all the bindings can
be put into the same module (algorithm) but the bindings themselves are present in individual
cxx files.

algorithm_def binding has been templated on implementation and trampoline  as it is required
for every algorithm