Be more forgiving when reading detected_object kpf
We'll only demand an ID, a Frame Number and a Bounding box
in each record (suffficent to build a detected_object).  We'll
more gently look for an Index (track id), detector_name and
confidence and use them if they're there.  Also update the
writer to write out detected_object->index.  This has the side
effect of allowing "GEOM" KPF annotation files from DIVA to
be read in as detected objects and provides enough information
(by adding index) to add annotation labels in a separate
process.