Update keyframe loop closure to store keyframes with the tracks.
Since this loop closure algorithm was written, the track set data structure
has been updated to stored frame-level data on the track set, including
a keyframe flag.  This commit updates the keyframe loop closure algorithm
to store the list of keyframes in the track set rather than internally.
This allows users of the tracks outside of this algorithm to know what
keyframes were selected.