Add metadata_item::is_valid
Add a new method to a metadata_item to test if it is valid or not. (The
"fake" item returned by metadata::find if the requested item is not
found will return false; all others will return true.) This makes it
possible to avoid having to look up items twice in order to use an item
if and only if it is actually present in the map.