Work around Visual Studio bug
Use size_t instead of int as the template type parameter when passing
around depth in the Qt image tests. This is necessary so that when the
parameter (via a class constexpr static member) gets passed to a
constructor that expects a size_t, the compiler doesn't have to
implement the exception of [dcl.init.list]/7.4, which Visual Studio
prior to VS2019 does not implement correctly.

See also https://developercommunity.visualstudio.com/content/problem/246960/msconnect-3131094-narrowing-conversion-from-int-co.html.