Fixed a bug in the Ceres bundle adjuster callback.
The main optimize function was previously updated to modify landmarks in place
for speed. This was fine, but the call back was not updated similarly.  As a
result, when the callback was used it would overwrite the landmarks with new
copies, but the main optimize function was not returning these changes since
it was returning the original memory.  This change updates the callback to
update in the same way as the main optimize function to fix this bug.