Fix use of range adapters on temporaries
Overhaul a bunch of the range guts to work correctly on temporaries, by
having the adapter take a copy of the input. (Note that this means that
range adapters can only be used on temporary inputs if the input
container type is movable, but with current lifetime rules, there is no
other way to lifetime-extend an input temporary.)

Actually getting this to work right, especially on GCC 4.8 (which has
some issues with overload resolution in the face of r-value references)
was... interesting. Hint: this, kids, is why std::forward exists!