Fix read_object_track_set_kw18
Modify read_object_track_set_kw18 so that, in batch mode (which is the
default?!), it doesn't violate the API by always returning the same
track set successfully, resulting in code that expects a generic
algorithm instance getting stuck in an infinite loop. Also clean up some
code in the same vicinity.