Explicitly require Qt ≥ 5.10
Modify finding Qt (when we use it) to explicitly require Qt ≥ 5.10. We
actually require this already, due to use of a QImage API that was added
in 5.10; this just arranges to get an explicit version-mismatch error at
configure time rather than a 'no such method' build error. (It would be
possible to work around this, at the cost of not supporting images over
2 GiB, but at least one of our consumers also requires Qt ≥ 5.10, and
fletch provides at least 5.11.)