Make detected_object's bounding box inline
Don't use a shared pointer to store the bounding box in detected_object;
instead, store it directly in the class. This is almost certainly less
expensive, since a bounding_box_d is a small, trivially copyable object,
and there was no way for the bounding box to not be set.