Don't pass scalars by reference
Tweak one of the constructors of bounding_box to take the width and
height by value, rather than by reference. These are almost certainly
scalars, which are trivial to copy by value, and this is consistent with
the constructor that takes each of {x,y}{min,max} individually, which
was already sensibly taking them by value.