Improve coding style documentation
Add a more comprehensive code style guide, along with explanations.
Modify CONTRIBUTING.rst to refer to the new guide.

Even if we had perfect styling tools (we don't, currently), a
comprehensive guide is useful both as an explanation of the reasons for
choosing a particular style, and for higher level guidelines that an
automated tool will not address (such as naming conventions and higher
level design principles).