Mark intrinsics as fixed if any camera using them is fixed.
Because cameras can have shared intrinsics we should not optimize
the shared intrinsics if some cameras using them are fixed.
Those fixed cameras will have residuals ignored in the optimization
because they should be fixed, but they will actually change because
a non-fixed camera may adjust the shared intrinsics.  This causes
bundle adjustment to increase error in fixed cameras.  The safe
alternative is to only optimize intrinsics if all cameras sharing
the intrinsics are fixed.