Removed limit on number of fixed cameras in BA
This feature seems like a hack that could come back to bite us.
I don't think it belongs here in this current form.  As written,
the code filters residual terms after encountering 4 fixed cameras.
The residual terms added depends on the order the cameras are stored,
which is not a good thing.  A better way to write this would be to
first check if there are at least N fixed cameras and then choose to
filter all or none of the residuals based on that.  However, I'm not
convinced that that version of this feature is really needed either.