Add names to the code style guide
Add section and rule names to the code style guide, roughly along the
lines of how the C++ standard likes to name things. This gives us a
mechanism (as requested by Aaron Bray) to refer to specific rules that
does not depend on numbers, which are more likely to change as the style
guide is revised.