Add altitude to the sensor location geopoint if found.
Since geo_point is now primarily a 3D point, if we have data on the
third coordinate it should be set.  It's less clear if we should also
set the altitude to its own metadata tag as was done historically.
For now the sensor altitude is set in both places.