Added a function to compute keyframes for SfM
These are not the same keyframes used by the tracker during tracking,
though it is likely that some of the same frames are chosen.  These frames
are explicitly selected to provide good initialization for SfM.  They aim
to be well distributed through the video but also well connected to each other.