Try to repair a fragmented camera graph
If the camera graph fractures into multiple connected components we were
previously discarding all but the largest component in hope that we could
rebuild it later.  In practice the discarded components can be large and
they do not get rebuilt.  This change tries to stitch the components back
together rather than discard them.  It finds the tracks that span across
the disjoint components and retriangulates them.  These points may be poorly
constrained, but they are required to hold the model together.