Added a second bundle adjuster instance to the initializer
The allows a separate "global" bundle adjuster instance that can be
configured differently.  The "global" BA instance is used for optimizing
all (or many) cameras at once that are already in good configuration.
The regular bundle adjuster is for correcting newly added cameras that
are less trusted.  With this change we can provided different optimization
paramters for each case.