Added a new parameter to depth integration for occluded space constant.
This new Epsilon value is used in the occluded portions of the volume
at a distance greater than Delta where the TSDF was previously defined to be
zero.  Using a very small positive value can be useful in preventing a
"shell" or "second surface" effect in which objects become hollow with
a thin shell because of the boundary between positive and zero values.
Setting Epsilon > 0 removes this second surface.

Epsilon > 0 is most useful when the depth viewpoint have coverage from all
angles.  When the depth viewpoint are limited, it may be better to set
Epsilon to 0.