Updated compute_depth callback implementation
The callback now takes a depth image, a status string, and
a percent complete integer (previously the integer was number of iterations).
It is now assumed that the callback may be called with progress information
only by leaving the depth image empty.  The Super3d implementations are
updated to use callbacks for both the cost volume construction (first 50% of
progress) and depth refinement (second 50% of progress).  Status updates
are more frequent than depth updates and allow for quicker early termination.