Aggressively prevent reparenting track states
Track states differ from most objects in that they back-reference their
owning track. Accordingly, we prevent them from being added to a track
if they already have an owning track. However, users may not be aware of
this caveat (and indeed, multiple instances of this form of misuse have
been recently uncovered), and may not be checking for failures to add
track states to a track, since the "obvious" failure cases can often be
statically known to not occur.

In order to help address these issues, modify the code which adds track
states to make it a logic error (i.e. literally throw std::logic_error)
when attempting to add an already-owned track state to a track, as this
almost always indicates that the code attempting to perform the
operation does, in fact, contain a logic error. This should make it much
easier to identify such code, rather than having to chase down more
subtle bugs due to track states unexpectedly "going missing".