Revert an earlier change to switch BA instances over time
Now that we switched to windowed bundle adjustment after establishing
keyframes, this hacky solution to reduce the number of BA iterations
when the solution gets large and near complete is no longer needed.
The windowed BAs tend to converge in just a few iterations anyway.
We don't want arbitrary thresholds on problem size to define when
to enforce fewer iterations.  Switching to windowed BA once keyframes
cameras are established does this implicitly in a better justified way.