Add time stamp, frame identifier to track writing
Modify the write_object_track_set algorithm API to also accept the frame
time stamp and frame identifier when writing the track set for a frame.
This is required for some formats which need to include this
information. Also modify write_object_track_process to pass along this
information to the algorithm.

Co-authored-by: Kyle Edwards <kyle.edwards@kitware.com>
Co-authored-by: Matt Dawkins <matt.dawkins@kitware.com>