Dev/add library dir to targets (#867)
* Support library_dir variable

* Add functions to manipulate library_dir

Added function to set library_dir to default when library_dir is not defined
Added function to remove leading and trailing slashes in library_dir

* Use library_dir variable instead of lib

* Better documentation of library_dir

* Generalize path validation

* Add helper function to compute reverse relative path for RPATH

As RPATH work with ${ORIGIN} we need a way to do compute reverse
relative path. Eg. for library_dir with value lib/foo the helper function would
return ../../. These relative path can be chained from library_dir and
library_subdir to navigate to project root

* Create KWIVER_LIBRARY_DIR to represent default library directory

* Use helper function in conjunction with KWIVER_LIBRARY_DIR

* Change KWIVER_LIBRARY_DIR to KWIVER_DEFAULT_LIBRARY_DIR

KWIVER_LIBRARY_DIR is already being used in kwiver-config-build and
kwiver-config-install. As the default library directory would have to
be export it is being renamed to KWIVER_DEFAULT_LIBRARY_DIR

* Export KWIVER_DEFAULT_LIBRARY_DIR