Make open and close virtual in detected_object_set_output
Due to limitations of C++ destructors, an overridden version of close
will not be called by detected_object_set_output's destructor, so
subclasses might need to call close in their own destructors.

The intent is to allow redefining this in Python implementors, which
means that the documented exceptions for open will not remain the
case.  (Though the list is currently incorrect at present, too, as
it's a kwiver::vital::file_not_found_exception that open can throw.)