Add shared flag to output port declarations.
These processes push an image container smart pointer to an
output port. This means that all processes that have connected
to this output port are sharing the same image. Sprokit provides
a way to declare this and check to make sure none of the
processes that are using this port will change the shared data.
This type of port usage compatibility is checked when the pipeline
is built.