[init] WIP avoid errors from reinitializations with the same ID
The main issue now is that we don't attempt to check that there's no
undesired overlap with an automatically generated ID.

This is difficult in part because some might be intentional.  Say the
tracker proposes its own track.  Will that then be used to initialize
a track in an earlier tracker?  If so, it might come back as an
explicit initialization for this tracker.