Implemented a faster alternative to counting number of frames in FFMPEG.
Instead of advancing through the entire video and counting every frame,
the new algorithm seeks to the last keyframe it can find and then advances
to the end to discover the frame number of the last frame.  With this change
the number of frames is accessible in constant time rather than linear in the
number of frames of video.