Improve sorting by time
Improve the functionality to sort video_input_image_list's input
according to the images' timestamps, as introduced in the previous
commit. In particular, improve the configuration description to avoid
reference to a particular image reader wrapper, and improve the
performance of the sorting algorithm. Also, don't mix up the input
argument to sort_by_time with the member m_files. (They are always the
same, but doing so still makes the code unnecessarily obtuse.)

Benchmarking has strongly indicated that building a flat list with the
path and timestamp pairs, then sorting with std::sort, is faster in all
cases than either using std::map or rolling and unrolling a heap.
Additionally, the algorithm has been tweaked so that the paths are never
copied (because we are rebuilding the file list, we can move the strings
out of the input list into the scratch list, then move them back), which
should significantly reduce overhead from memory allocation. This
approach should require only a single allocation for the temporary
vector which will be sorted, which likely explains why it outperforms a
std::map which requires multiple node allocations. (It almost surely
doesn't hurt either to defer sorting until the entire list is complete
rather than sorting on every insertion.)