Add a basic signal/slot mechanism
Add a simple version of a signal class which can be used to notify
observers of events. Add a context to manage slot connections to the
same. For now, an explicit context is required, and manual disconnection
of slots is not supported (this could be added in the future). Also,
concurrently emitting a signal from multiple threads is not supported,
as doing so requires a read/write lock, which we do not currently have
(std::shared_mutex does not exist until C++17).