Add signal for detected_object_type type names
Add a signal, using the machinery introduced in the previous commit,
which is emitted when a new type name is added to the type name list
managed by detected_object_type. This is useful if users need to execute
some logic when this occurs, e.g. for an application that needs to
persistently display all known type names.