Implement add_copy() method to take shared pointer.
An additional add_copy() method was needed to support python bindings. Python
manages objects with shared pointers and this collection did not support
that type of reference. An additional complication is that the metadata
collection manages its own elements and can not directly accept an element
managed by a shared_ptr due to the design of the collection. The metadata
item managed by the shared pointer is copied and the copy is added to
the collection. The copy is low overhead since metadata objects are
small, and metadata collections are not created that frequently.

Co-authored-by: Matthew Woehlke <matthew.woehlke@kitware.com>