Improve metadata API
Modify metadata_map::add to take a unique_ptr rather than a raw pointer to
make it clear that the map takes ownership of the item. Add overloads to
create the item, replacing the ugly NEW_METADATA_ITEM macro. Replace
uses of said macro with the new methods.

This is also somewhat more type-safe, as the new method requires that
the data being added is of the correct type. (An add_any method is
provided for users that need to be able to add items without static type
checking.)