Fixed a bug causing early termination of the camera init loop
The camera initialization loop is supposed to process all keyframes first
and when those are used up it processes the rest of the frames.  There
was a bug in catching this transition.  If the last keyframe camera failed
to initialized the loop would exit before switching to process the remaining
cameras.  The switching code has not been moved to the start of the loop
rather than the end to avoid this.