Reseat arrows/pytorch to sprokit/processes/pytorch
The build destination has been changed for none of the files.  As the
processes directory's inclusion is already gated by
KWIVER_ENABLE_PROCESSES, we drop it from pytorch/CMakeLists.txt.

This theoretically results in the loss of the ability to use the
support modules without enabling process support, but note that said
modules are not arrows in the KWIVER sense, so it's not clear where
they would belong outside of this anyway.