Remove ${CMAKE_INSTALL_PREFIX} from destination paths.
The addition of this prefix was breaking some use of kwiver_add_library
in external projects that build on KWIVER.  Some CPack packagers require
these paths to be relative.  When absolute paths are needed, CMake will
automatically prepend CMAKE_INSTALL_PREFIX, so there should be no
need to make this explicit.