Add recommendations for commits to CONTRIBUTING
Add a short section on best practices for commits/history and commit
messages to CONTRIBUTING.rst. Note that the goal here isn't to beat
people over the head with these policies, but rather to have a statement
somewhere that we can easily reference as to our preferences and
additional resources.