Update KWIVER to include more fixes
This includes a workaround for failed inter-camera registration at the
start of a sequence.  (Just say the camera has the same homography as
its neighbor.)

It also includes shortening the name of
augmented_resnet_feature_extractor to just augment_feature_extractor
in order to keep the Windows build path lengths short.