Use Eigen3's CMake package
Remove our find module for Eigen and suppress use of the one that Eigen
itself ships, instead requiring the CMake package configuration. Use
Eigen's exported targets (which only exist using the package
configuration rather than the find module) instead of global includes.

This gets us a little closer to consumers being able to use KWIVER
without having to hack around our prior failure to export dependency on
Eigen. (Having public target includes instead of global includes would
be even easier, as this configuration currently requires that consumers
first find Eigen themselves, but that approach also breaks down if
consumers need to use Eigen from a different location. The proper fix is
for KWIVER's own package configuration to require Eigen, with a hint
where we expect to find it, but that's currently painful to do
correctly. This may be addressed in a future commit.)