Add support for OpenCV 4.x
Update OpenCV arrow to be able to compile against OpenCV 4.x. A shocking
lot of the changes are simply "don't rely on the C API", which we
shouldn't be (ab)using in C++ code anyway. In fact, just stop including
any C API headers, which seems to be much more portable across different
versions of OpenCV anyway.