Rename OpenCV version test symbols for clarity
Replace KWIVER_HAS_OPENCV_VER_{3,4} version test preprocessor symbols
with KWIVER_OPENCV_VERSION_MAJOR, which is defined to the major version
number of OpenCV (cuyrrently, `2`, `3` or `4`). This makes it clearer
when we have code that needs 3.x or newer, rather than the old setup
which implicitly (and perhaps confusingly) used the symbols to mean 'or
newer'.