Add WIX capability to VIAME
Add a new section of build_package.cmake which sets and uses the WIX
toolset to build a Windows executable.

Prevent the system libraries from being written to prevent duplications
of the files in the installer.

When packaging is selected, force off "VIAME_DOWNLOAD_MODELS" as Wix has
a file size limit of 2GB for data.