"Apply" uncrustify to arrows/qt
Run uncrustify on arrows/qt and apply those changes that make sense.
This excludes a modest number of changes that should not have been made
and represent either bugs, configuration issues, or lack of desired
features. (For the latter, particularly, uncrustify currently does not
have a way to add blank lines between the start/end of a namespace and
content thereof while excluding blank lines in between different levels
of namespaces. On the other hand, I'm not sure if we're going to try to
implement this, or at some point just accept having blank lines there as
well, but for now, don't add them.)