Fixed a bundle_adjust parameter that was manually overridden.
The ceres bundle adjuster has an option update_state_every_iteration
that controls whether the state is updated on each iteration, generally
for the purpose of returning in a callback.  Previously we set this
option based on whether a callback was registered or not.  However, we
have since exposed this option in the algorithm config block.  There are
cases where we want a callback (e.g. to terminate early) but not updated
state.  However, we previously forgot to remove the old code that sets
this open based on whether a callback was set.  So the config option was
always overridden.