Fix partial loading of plugins
Add flags type to vital. Use this to specify types of plugins to load;
essentially this gives us nice operators for enumerators meant to be
used as bitmasks, which greatly improves user convenience. Tweak loading
to keep track of which plugins are loaded, so that multiple requests to
load different types of plugins works as expected.