Fix high-pass filter for certain pixel types
Tweak the VXL high-pass filter to use a pixel-type appropriate type for
the accumulator instead of always using unsigned, which will work quite
badly for signed or floating point pixel types.

Note that this drops support for bitmaps (pixel type BOOL), for which
smoothing is a rather questionable operation anyway, and for 64-bit
integer pixel types, as these would require using a floating-point type
for the accumulator, which would need casts to avoid compiler warnings
and could result in non-trivial loss of precision. (The latter probably
is not a huge issue, but for now, it's unlikely we actually need to
operate on such images, anyway.)