Don't remove newlines after template
Change nl_template_func_decl in the uncrustify configuration from remove
to ignore. For multi-line declarations, we prefer to break after the
template argument list (i.e. add a newline here, not remove it!),
however we also want to accept one-line declarations. As there is no way
currently to enforce removal only if it results in a one-liner, just
leave it alone for now.