Add "copy" option to vxl_convert_image
Add an additional value to the format option of vxl_convert_image to
copy the input format. This differs slightly from the existing
"disable", which makes the filter a no-op, in that while no value type
conversion is applied, other operations such as scaling or converting to
single channel may still be applied.