Work around broken MSVC
Replace explicit lambda captures in the scale_image helper function of
vxl_convert_image with a default-capture-by-value specification. This is
needed to work around a bug in the MSVC compiler that conflicts with our
previous attempt (5fa46f9c338) to fix a warning about an unused capture.

The C++11 standard (n.b. the example at [expr.prim.lambda]¶12) is pretty
clear that capturing a constant from the containing scope is NOT needed,
however it seems that as of writing, a full decade later, no version of
MSVC implements this correctly.