Fixed flipped and transposed depth image memory
The code for loading depth images from a VTK file was storing the data both
flipped horizontally and transposed in the vital image.  The only user of
this data, CUDA depth fusion, was then transposing again during copy to
CUDA memory and flipping during pixel access to compensate.  This worked,
but any other algorithm using the same images would see the same garbled
memory and need would need to do the same operations to make sense of it.

This change removes the flip and transpose in both locations so that the
image passed in between is valid and can be used for other purposes.