Make end of pipeline detectable
Add EmbeddedPipelineWorker::finished signal to notify users when the
output endcap has received an end-of-input notification. This allows
better handling of pipeline completion for users that aren't tracking
completion based on number of frames processed. (This is also beneficial
to detect when the pipeline has halted if input is terminated
prematurely, as when aborting pipeline execution.)