Add KWIVER_CONFIG_PREFIX environment variable
Modify config search path logic to accept an environment variable
KWIVER_CONFIG_PREFIX as an addition to the normal KWIVER install prefix.
This provides a convenient mechanism for developers to ensure that
shared data from a build tree can be located. (This is more convenient
than KWIVER_CONFIG_PATH since our shared data is by default installed
to a version-specific location. Using KWIVER_CONFIG_PREFIX allows
pointing directly to the root of the build tree and does not need to be
changed when the KWIVER version changes.)

While we're at it, slightly streamline the code for generating the
search paths.