Fix track_features API
Alter track_features::track to take the frame as a frame_id_t rather
than as unsigned. Ironically, this isn't just "more correct", but
algorithm implementations were already correctly operating with frames
as frame_id_t.

Also, tweak implementations to use the override keyword, which helps
ensure that the overrides are in fact the correct API, and will lead to
more direct errors ('marked override but does not override' when parsing
the class definition, rather than failures to instantiate the class due
to abstract virtuals) in case something changes again.