Fix handling of corrective reinitializations
If a last-frame reinitialization is given for a track that already had
an update on the last frame, the track currently ends up with two
states for the same frame.  This teaches track.append a parameter
"on_duplicate" which can take a value "replace" that replaces the
first state.

The various possible ValueErrors here should never be raised in
current use.  In particular, trying to replace just after clearing
error should never happen because that would require using
track_set.make_track with on_exist="restart" and then appending a
previous-frame initialization, which isn't done.