Remove assertion on previous-frame initializations
This checked that initializations coming in on different frames but
sharing a track ID and frame number were in fact the same detection.
This check doesn't make sense in combination with the track conductor,
where initializations may either come from the track conductor and be
from the previous frame or come from an external source and probably
be from the current frame.  We might expect the detections to be
equivalent, but there's no guarantee they're literally the same
object.

With this check gone, the resulting behavior is that the
previous-frame initialization is ignored, which is the desired
behavior on the assumption that the two detections are effectively the
same.